+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+- Add contextual bindings via GacelaConfig::when()
+- Add service aliases via GacelaConfig::addAlias()
+- Add protected services via GacelaConfig::addProtected()
+
 ## [1.12.0](https://github.com/gacela-project/gacela/compare/1.11.0...1.12.0) - 2025-11-09
 
 - Renamed `DocBlockResolver` to `ServiceResolver` to better reflect its purpose